refactor: 선택한 이미지 미리보기 위젯 분리
